Move exec batch logic from flow executor to executor proxy
