+++ src/promptflow/promptflow/batch/_python_executor_proxy.py
@@ -6,11 +6,13 @@ from pathlib import Path
 from typing import Any, List, Mapping, Optional
 
 from promptflow._constants import LINE_TIMEOUT_SEC
+from promptflow._core._errors import UnexpectedError
 from promptflow._core.operation_context import OperationContext
 from promptflow.batch._base_executor_proxy import AbstractExecutorProxy
 from promptflow.contracts.run_mode import RunMode
 from promptflow.executor import FlowExecutor
 from promptflow.executor._flow_nodes_scheduler import DEFAULT_CONCURRENCY_BULK
+from promptflow.executor._line_execution_process_pool import LineExecutionProcessPool
 from promptflow.executor._result import AggregationResult, LineResult
 from promptflow.storage._run_storage import AbstractRunStorage
 
@@ -59,12 +61,36 @@ class PythonExecutorProxy(AbstractExecutorProxy):
         self._flow_executor._node_concurrency = DEFAULT_CONCURRENCY_BULK
         with self._flow_executor._run_tracker.node_log_manager:
             OperationContext.get_instance().run_mode = RunMode.Batch.name
-            line_results = self._flow_executor._exec_batch_with_process_pool(
-                batch_inputs, run_id, output_dir, validate_inputs=True, batch_timeout_sec=batch_timeout_sec
+            line_results = self._exec_batch_with_process_pool(
+                batch_inputs, output_dir, run_id, batch_timeout_sec=batch_timeout_sec
             )
             # For bulk run, currently we need to add line results to run_tracker
             self._flow_executor._add_line_results(line_results)
         return line_results
+
+    def _exec_batch_with_process_pool(
+        self,
+        batch_inputs: List[Mapping[str, Any]],
+        output_dir: Path,
+        run_id: Optional[str] = None,
+        batch_timeout_sec: Optional[int] = None,
+    ) -> List[LineResult]:
+        nlines = len(batch_inputs)
+        line_number = [batch_input["line_number"] for batch_input in batch_inputs]
+
+        if self._flow_executor._flow_file is None:
+            raise UnexpectedError(
+                "Unexpected error occurred while init FlowExecutor. Error details: flow file is missing."
+            )
+
+        with LineExecutionProcessPool(
+            self._flow_executor,
+            nlines,
+            run_id,
+            output_dir,
+            batch_timeout_sec=batch_timeout_sec,
+        ) as pool:
+            return pool.run(zip(line_number, batch_inputs))
 
     @classmethod
     def _get_tool_metadata(cls, flow_file: Path, working_dir: Path) -> dict:
